+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/PrettyCore.hs
@@ -1006,14 +1006,13 @@ plutusSummary (Alonzo _) (TimelockScript x) = timelockSummary x
 plutusSummary other _ = ppString ("Plutus script in era " ++ show other ++ "???")
 
 dStateSummary :: DState c -> PDoc
-dStateSummary (DState umap future (GenDelegs current) irwd deposit) =
+dStateSummary (DState umap future (GenDelegs current) irwd) =
   ppRecord
     "DState"
     [ ("Unified Reward Map", uMapSummary umap),
       ("Future genesis key delegations", ppInt (Map.size future)),
       ("Genesis key delegations", ppInt (Map.size current)),
-      ("Instantaneous Rewards", instantSummary irwd),
-      ("deposit", ppInt (Map.size deposit))
+      ("Instantaneous Rewards", instantSummary irwd)
     ]
 
 instantSummary :: InstantaneousRewards c -> PDoc