+++ libs/ledger-state/src/Cardano/Ledger/State/Query.hs
@@ -21,6 +21,7 @@ import Cardano.Ledger.State.UTxO
 import Cardano.Ledger.State.Vector
 import qualified Cardano.Ledger.TxIn as TxIn
 import Cardano.Ledger.UMapCompact (delView, ptrView, rewView, unify)
+import qualified Cardano.Ledger.UMapCompact as UM
 import qualified Cardano.Ledger.UTxO as Shelley
 import Conduit
 import Control.Foldl (Fold (..))
@@ -530,7 +531,8 @@ getDStateNoSharing dstateId = do
       rws <- selectList [RewardDstateId ==. dstateId] []
       forM rws $ \(Entity _ Reward {..}) -> do
         Credential credential <- getJust rewardCredentialId
-        pure (Keys.coerceKeyRole credential, rewardCoin)
+        pure (Keys.coerceKeyRole credential, UM.RDPair (UM.compactCoinOrError rewardCoin) (UM.CompactCoin 0))
+  -- FIXME the deposit is not accounted for ^
   delegations <-
     Map.fromList <$> do
       ds <- selectList [DelegationDstateId ==. dstateId] []
@@ -567,8 +569,7 @@ getDStateNoSharing dstateId = do
               iRTreasury = iRTreasury,
               deltaReserves = dStateIrDeltaReserves,
               deltaTreasury = dStateIrDeltaTreasury
-            },
-        dsDeposits = Map.empty -- FIXME, HELP ME FIX THIS
+            }
       }
 
 getDStateWithSharing ::
@@ -580,7 +581,8 @@ getDStateWithSharing dstateId = do
       rws <- selectList [RewardDstateId ==. dstateId] []
       forM rws $ \(Entity _ Reward {..}) -> do
         Credential credential <- getJust rewardCredentialId
-        pure (Keys.coerceKeyRole credential, rewardCoin)
+        pure (Keys.coerceKeyRole credential, UM.RDPair (UM.compactCoinOrError rewardCoin) (UM.CompactCoin 0))
+  -- FIXME the deposit is not accounted for ^
   delegations <-
     Map.fromList <$> do
       ds <- selectList [DelegationDstateId ==. dstateId] []
@@ -621,8 +623,7 @@ getDStateWithSharing dstateId = do
               iRTreasury = iRTreasury,
               deltaReserves = dStateIrDeltaReserves,
               deltaTreasury = dStateIrDeltaTreasury
-            },
-        dsDeposits = Map.empty -- FIXME, HELP ME FIX THIS TOO
+            }
       }
 
 loadDStateNoSharing :: MonadUnliftIO m => T.Text -> m (Shelley.DState C)